requirements.txt
motor<=3.7.0
python-dotenv<=0.21.0
git+https://github.com/Rapptz/discord-ext-menus
discord.py[speed]
uvloop
